menambahkan relasi didata buku ke nama pengarang dan ke nama penerbit pada tabel penerbit dan pengarang

+++ database/migrations/2023_04_28_055827_create_bukus_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('judul', 100);
             $table->string('pengarang', 100);
             $table->string('penerbit', 100);
-            $table->bigInteger('genre_id')->index();
+            $table->BigInteger('genre_id')->index();
             $table->string('tahun_terbit', 100);
             $table->string('gambar', 225)->nullable();
             $table->timestamps();
